[step2] refactor : 패키지 분리, 총 수익률계산 메서드 수정, 상수 사용
