+++ src/main/java/step2/domain/LottoBalls.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 import java.util.List;
 import java.util.stream.Collectors;
