+++ src/main/java/step2/domain/LottoTicketGenerator.java
@@ -1,4 +1,4 @@
-package step2;
+package step2.domain;
 
 import java.util.Collections;
 import java.util.List;
