+++ src/main/java/step2/view/InputView.java
@@ -16,8 +16,10 @@ public class InputView {
 
     public static List<Integer> askLuckyNumber() {
         printBlackLine();
+        scanner.nextLine();
         System.out.println("지난 주 당첨 번호를 입력해주세요.");
         String inputLuckyNumber = scanner.nextLine();
+        printBlackLine();
 
         return Arrays.stream(inputLuckyNumber.split(LUCKY_NUMBER_SEPARATOR))
                 .map(String::trim)
