+++ src/main/java/step2/view/OutputView.java
@@ -1,7 +1,7 @@
 package step2.view;
 
-import step2.LottoRank;
-import step2.LottoTicket;
+import step2.domain.LottoRank;
+import step2.domain.LottoTicket;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -9,32 +9,35 @@ import java.util.Map;
 
 public class OutputView {
 
+    public static final int PRICE_FOR_A_LOTTO_TICKET = 1000;
+    public static final String MESSAGE_FOR_EARNING_RATE = "총 수익률은 %.2f 입니다.";
+    public static final String MESSAGE_FOR_LOTTO_RESULT = "%d개 일치 (%d원) - %d개";
+
     public static void printNumberOfLottoTickets(int numberOfLottoTicketsToBuy) {
         System.out.println(numberOfLottoTicketsToBuy + "개를 구매했습니다.");
     }
 
     public static void printLottoTickets(List<LottoTicket> lottoTickets) {
-
-                lottoTickets.stream()
-                    .map(LottoTicket::getLottoTicket)
-                    .forEach(System.out::println);
-
+        lottoTickets.stream()
+                .map(LottoTicket::getLottoTicket)
+                .forEach(System.out::println);
     }
 
     public static void printResult(Map<LottoRank, List<LottoTicket>> lottoResult, int numberOfTotalTickets) {
         System.out.println("당첨 통계");
         System.out.println("---------");
-        int totalPrizeMoney = 0;
+        long totalPrizeMoney = 0;
         for (LottoRank rank : LottoRank.values()) {
-            totalPrizeMoney += rank.getPrizeMoney();
-            System.out.println(String.format("%d개 일치 (%d원) - %d개",
+            System.out.println(String.format(MESSAGE_FOR_LOTTO_RESULT,
                     rank.getNumberOfMatchedToLuckyNumber(),
                     rank.getPrizeMoney(),
                     lottoResult.getOrDefault(rank, new ArrayList<>()).size()));
+
+            totalPrizeMoney += rank.getPrizeMoney() * lottoResult.getOrDefault(rank, new ArrayList<>()).size();
         }
 
-        System.out.println(String.format("총 수익률은 %f 입니다.",
-                (double) totalPrizeMoney / numberOfTotalTickets * 1000));
+        System.out.println(String.format(MESSAGE_FOR_EARNING_RATE,
+                (double) totalPrizeMoney / (numberOfTotalTickets * PRICE_FOR_A_LOTTO_TICKET)));
     }
 
 }
