+++ src/test/java/step2/LottoTicketTest.java
@@ -1,6 +1,8 @@
 package step2;
 
 import org.junit.jupiter.api.Test;
+import step2.domain.LottoNumber;
+import step2.domain.LottoTicket;
 
 import java.util.Arrays;
 import java.util.List;
